src/requirements.txt
streamlit
opencv-python-headless
torch
torchvision
scipy
pandas
numpy
Pillow